requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.1
joblib==1.4.2
firebase-admin==6.5.0
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
